Remove Go -longsize option.

+++ Examples/test-suite/go/template_opaque_runme.go
@@ -3,7 +3,7 @@ package main
 import "./template_opaque"
 
 func main() {
-	v := template_opaque.NewOpaqueVectorType(10)
+	v := template_opaque.NewOpaqueVectorType(int64(10))
 
 	template_opaque.FillVector(v)
 }
